+++ include/envp.h
@@ -1,0 +1,6 @@
+#ifndef ENVP_H
+# define ENVP_H
+
+#include "types/t_envp.h"
+
+#endif
