+++ include/types/t_envp.h
@@ -1,0 +1,14 @@
+#ifndef ENVP_H
+# define ENVP_H
+
+typedef struct s_envp t_envp;
+
+struct s_envp
+{
+	char	*key;
+	char	*value;
+	t_envp	*next;
+	int		cnt;
+};
+
+#endif
